feat: implement MonitoringValue repository and refactor related use cases
