+++ src/compositionRoot.ts
@@ -109,7 +109,6 @@ import { UpdateMonitoringUseCase } from "./domain/reports/mal-data-approval/usec
 import { UserGroupD2Repository } from "./data/reports/mal-data-approval/UserGroupD2Repository";
 import { MonitoringValueDataStoreRepository } from "./data/reports/mal-data-approval/MonitoringValueDataStoreRepository";
 import { CountryCodeD2Repository } from "./data/reports/mal-data-approval/CountryCodeD2Repository";
-import { GetMonitoringValueUseCase } from "./domain/reports/mal-data-approval/usecases/GetMonitoringValueUseCase";
 
 export function getCompositionRoot(api: D2Api) {
     const configRepository = new Dhis2ConfigRepository(api, getReportType());
@@ -167,8 +166,11 @@ export function getCompositionRoot(api: D2Api) {
             getColumns: new GetMalDataApprovalColumnsUseCase(dataDuplicationRepository),
             saveColumns: new SaveMalDataApprovalColumnsUseCase(dataDuplicationRepository),
             getMonitoring: new GetMonitoringUseCase(monitoringValueRepository),
-            getMonitoringValue: new GetMonitoringValueUseCase(countryCodeRepository, userGroupRepository),
-            updateMonitoring: new UpdateMonitoringUseCase(monitoringValueRepository, userGroupRepository),
+            updateMonitoring: new UpdateMonitoringUseCase(
+                monitoringValueRepository,
+                countryCodeRepository,
+                userGroupRepository
+            ),
             updateStatus: new UpdateMalApprovalStatusUseCase(
                 dataDuplicationRepository,
                 dataValuesRepository,
