+++ src/domain/reports/mal-data-approval/usecases/UpdateMonitoringUseCase.ts
@@ -1,9 +1,11 @@
+import _ from "lodash";
 import { UseCase } from "../../../../compositionRoot";
 import { MalDataApprovalItemIdentifier, MalDataSet } from "../entities/MalDataApprovalItem";
 import { UserGroupRepository } from "../repositories/UserGroupRepository";
+import { CountryCodeRepository } from "../repositories/CountryCodeRepository";
+import { CountryCode } from "../entities/CountryCode";
 import { MonitoringValueRepository } from "../repositories/MonitoringValueRepository";
 import { Monitoring, MonitoringValue } from "../entities/MonitoringValue";
-import _ from "lodash";
 
 type UpdateMonitoringUseCaseOptions = {
     namespace: string;
@@ -16,6 +18,7 @@ type UpdateMonitoringUseCaseOptions = {
 export class UpdateMonitoringUseCase implements UseCase {
     constructor(
         private monitoringValueRepository: MonitoringValueRepository,
+        private countryCodeRepository: CountryCodeRepository,
         private userGroupRepository: UserGroupRepository
     ) {}
 
@@ -25,15 +28,17 @@ export class UpdateMonitoringUseCase implements UseCase {
         const dataNotificationsUserGroup = await this.userGroupRepository.getUserGroupByCode(
             malDataNotificationsUserGroup
         );
-        const updatedMonitoringValue = buildMonitoringValue(
+        const countryCodes = await this.countryCodeRepository.getCountryCodes();
+        const monitoring = buildMonitoringValue(
             monitoringValue,
             dataApprovalItems,
             dataSetName,
+            countryCodes,
             dataNotificationsUserGroup.id,
             enableMonitoring
         );
 
-        return this.monitoringValueRepository.save(namespace, updatedMonitoringValue);
+        return this.monitoringValueRepository.save(namespace, monitoring);
     }
 }
 
@@ -41,60 +46,57 @@ function buildMonitoringValue(
     monitoringValue: MonitoringValue,
     dataApprovalItems: MalDataApprovalItemIdentifier[],
     dataSetName: MalDataSet,
+    countryCodes: CountryCode[],
     dataNotificationsUserGroup: string,
     enableMonitoring: boolean
 ): MonitoringValue {
-    const initialDatasetMonitoring = _.first(monitoringValue.dataSets[dataSetName]);
-
-    if (!initialDatasetMonitoring) {
+    const addedMonitoringValues = dataApprovalItems.map(item => {
         return {
-            ...monitoringValue,
-            dataSets: {
-                ...monitoringValue.dataSets,
-                [dataSetName]: [
-                    {
-                        monitoring: dataApprovalItems.map(dataApprovalItem => ({
-                            orgUnit: dataApprovalItem.orgUnitCode,
-                            period: dataApprovalItem.period,
-                            enable: enableMonitoring,
-                        })),
-                        userGroups: [dataNotificationsUserGroup],
-                    },
-                ],
-            },
-        };
-    } else {
-        const initialMonitoring = initialDatasetMonitoring.monitoring;
-        const addedMonitoring = dataApprovalItems.map(dataApprovalItem => ({
+            orgUnit: item.orgUnit,
+            period: item.period,
             enable: enableMonitoring,
-            period: dataApprovalItem.period,
-            orgUnit: dataApprovalItem.orgUnitCode,
-        }));
-
-        const mergedMonitoringData = [...initialMonitoring, ...addedMonitoring].reduce<Record<string, Monitoring>>(
-            (acc, monitoringItem) => ({
-                ...acc,
-                [`${monitoringItem.orgUnit}-${monitoringItem.period}`]: {
-                    ...acc[`${monitoringItem.orgUnit}-${monitoringItem.period}`],
-                    ...monitoringItem,
-                },
-            }),
-            {}
-        );
-
-        return {
-            ...monitoringValue,
-            dataSets: {
-                ...monitoringValue.dataSets,
-                [dataSetName]: [
-                    {
-                        monitoring: Object.values(mergedMonitoringData),
-                        userGroups: [dataNotificationsUserGroup],
-                    },
-                ],
-            },
         };
-    }
+    });
+
+    const initialMonitoring = _.first(monitoringValue["dataSets"]?.[dataSetName])?.monitoring ?? [];
+    const newDataSets = _.merge({}, monitoringValue["dataSets"], {
+        [dataSetName]: [
+            _.omit(
+                {
+                    monitoring: combineMonitoringValues(initialMonitoring, addedMonitoringValues).map(monitoring => {
+                        return {
+                            ...monitoring,
+                            orgUnit:
+                                monitoring.orgUnit.length > 3
+                                    ? countryCodes.find(countryCode => countryCode.id === monitoring.orgUnit)?.code
+                                    : monitoring.orgUnit,
+                        };
+                    }),
+                    userGroups: [dataNotificationsUserGroup],
+                },
+                "userGroup"
+            ),
+        ],
+    });
+
+    return {
+        ...monitoringValue,
+        dataSets: newDataSets,
+    };
+}
+
+function combineMonitoringValues(
+    initialMonitoringValues: Monitoring[],
+    addedMonitoringValues: Monitoring[]
+): Monitoring[] {
+    const combinedMonitoringValues = addedMonitoringValues.map(added => {
+        return initialMonitoringValues.filter(
+            initial => initial.orgUnit !== added.orgUnit || initial.period !== added.period
+        );
+    });
+    const combinedMonitoring = _.union(_.intersection(...combinedMonitoringValues), addedMonitoringValues);
+
+    return _.union(combinedMonitoring);
 }
 
 const malDataNotificationsUserGroup = "MAL_Data_Notifications";
